requirements.txt
wavelink==3.4.1
discord.py==2.4.0
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.4
python-dotenv==1.0.0
pyodbc==4.0.39